Rename Modal => LegacyModal
