+++ src/components/SettingsPage/ResetButton.js
@@ -8,7 +8,7 @@ import type { T } from 'types/common'
 import hardReset from 'helpers/hardReset'
 import Box from 'components/base/Box'
 import Button from 'components/base/Button'
-import { ConfirmModal } from 'components/base/Modal'
+import { ConfirmModal } from 'components/base/LegacyModal'
 import IconTriangleWarning from 'icons/TriangleWarning'
 
 type Props = {
