+++ src/components/modals/AccountSettingRenderBody.js
@@ -34,7 +34,7 @@ import {
   ModalFooter,
   ModalContent,
   ConfirmModal,
-} from 'components/base/Modal'
+} from 'components/base/LegacyModal'
 
 type State = {
   accountName: ?string,
