+++ src/components/modals/Debug.js
@@ -5,7 +5,7 @@ import { connect } from 'react-redux'
 import { createStructuredSelector } from 'reselect'
 import { getCryptoCurrencyById } from '@ledgerhq/live-common/lib/helpers/currencies'
 import last from 'lodash/last'
-import Modal, { ModalBody, ModalTitle, ModalContent } from 'components/base/Modal'
+import Modal, { ModalBody, ModalTitle, ModalContent } from 'components/base/LegacyModal'
 import { getCurrentDevice } from 'reducers/devices'
 import Button from 'components/base/Button'
 import Box from 'components/base/Box'
